MAINT: Move expected rate lookup functions into fixture for reuse.
